+++ src/app/(auth)/kakao/callback/page.tsx
@@ -22,7 +22,11 @@ const KakaoCallback = () => {
       try {
         const user = await postKakaoLogin(code);
         login(user); // Zustand 상태 업데이트
-        router.replace("/"); // 로그인 성공 후 메인 페이지로 이동
+        
+        const redirectPath = localStorage.getItem("redirect_after_login") || "/";
+        localStorage.removeItem("redirect_after_login");
+        router.replace(redirectPath); // 원래 페이지로 이동
+
       } catch (error) {
         console.error("카카오 로그인 실패:", error);
         router.push("/"); // 실패 시 메인 페이지로 이동
